Refactor task operations
